Introduce MediaDeviceMonitor trait object
