+++ backends/dummy/lib.rs
@@ -9,7 +9,7 @@ extern crate servo_media_webrtc;
 
 use boxfnonce::SendBoxFnOnce;
 use ipc_channel::ipc::IpcSender;
-use servo_media::{Backend, BackendInit, MediaDeviceInfo, SupportsMediaType};
+use servo_media::{Backend, BackendInit, SupportsMediaType};
 use servo_media_audio::block::{Block, Chunk};
 use servo_media_audio::context::{AudioContext, AudioContextOptions};
 use servo_media_audio::decoder::{AudioDecoder, AudioDecoderCallbacks, AudioDecoderOptions};
@@ -19,6 +19,7 @@ use servo_media_audio::{AudioBackend, AudioStreamReader};
 use servo_media_player::context::PlayerGLContext;
 use servo_media_player::{audio, video, Player, PlayerError, PlayerEvent, StreamType};
 use servo_media_streams::capture::MediaTrackConstraintSet;
+use servo_media_streams::device_monitor::{MediaDeviceInfo, MediaDeviceMonitor};
 use servo_media_streams::registry::{register_stream, unregister_stream, MediaStreamId};
 use servo_media_streams::{MediaOutput, MediaSocket, MediaStream, MediaStreamType};
 use servo_media_traits::{ClientContextId, MediaInstance};
@@ -114,8 +115,8 @@ impl Backend for DummyBackend {
         SupportsMediaType::No
     }
 
-    fn enumerate_devices(&self) -> Result<Vec<MediaDeviceInfo>, ()> {
-        Ok(Vec::new())
+    fn get_device_monitor(&self) -> Box<dyn MediaDeviceMonitor> {
+        Box::new(DummyMediaDeviceMonitor {})
     }
 }
 
@@ -344,5 +345,13 @@ impl MediaInstance for DummyPlayer {
 
     fn resume(&self) -> Result<(), ()> {
         Ok(())
+    }
+}
+
+struct DummyMediaDeviceMonitor;
+
+impl MediaDeviceMonitor for DummyMediaDeviceMonitor {
+    fn enumerate_devices(&self) -> Result<Vec<MediaDeviceInfo>, ()> {
+        Ok(vec![])
     }
 }
