+++ backends/gstreamer/device_monitor.rs
@@ -1,24 +1,15 @@
-use atomic_refcell::AtomicRefCell;
 use gst::Caps;
 use gst::DeviceExt;
 use gst::DeviceMonitor as GstDeviceMonitor;
 use gst::DeviceMonitorExt;
 use gst::DeviceMonitorExtManual;
 
-use servo_media::{MediaDeviceInfo, MediaDeviceKind};
+use servo_media_streams::device_monitor::{MediaDeviceInfo, MediaDeviceKind, MediaDeviceMonitor};
 
-pub struct GStreamerDeviceMonitor {
-    devices: AtomicRefCell<Option<Vec<MediaDeviceInfo>>>,
-}
+pub struct GStreamerDeviceMonitor;
 
-impl GStreamerDeviceMonitor {
-    pub fn new() -> GStreamerDeviceMonitor {
-        GStreamerDeviceMonitor {
-            devices: AtomicRefCell::new(None),
-        }
-    }
-
-    fn get_devices(&self) -> Result<Vec<MediaDeviceInfo>, ()> {
+impl MediaDeviceMonitor for GStreamerDeviceMonitor {
+    fn enumerate_devices(&self) -> Result<Vec<MediaDeviceInfo>, ()> {
         const AUDIO_SOURCE: &str = "Audio/Source";
         const AUDIO_SINK: &str = "Audio/Sink";
         const VIDEO_SOURCE: &str = "Video/Source";
@@ -31,31 +22,20 @@ impl GStreamerDeviceMonitor {
         let devices = device_monitor
             .get_devices()
             .iter()
-            .map(|device| {
+            .filter_map(|device| {
                 let display_name = device.get_display_name().as_str().to_owned();
-                MediaDeviceInfo {
+                Some(MediaDeviceInfo {
                     device_id: display_name.clone(),
                     kind: match device.get_device_class().as_str() {
                         AUDIO_SOURCE => MediaDeviceKind::AudioInput,
                         AUDIO_SINK => MediaDeviceKind::AudioOutput,
                         VIDEO_SOURCE => MediaDeviceKind::VideoInput,
-                        _ => MediaDeviceKind::__Unknown,
+                        _ => return None,
                     },
                     label: display_name,
-                }
+                })
             })
             .collect();
-        Ok(devices)
-    }
-
-    pub fn enumerate_devices(&self) -> Result<Vec<MediaDeviceInfo>, ()> {
-        {
-            if let Some(ref devices) = *self.devices.borrow() {
-                return Ok(devices.clone());
-            }
-        }
-        let devices = self.get_devices()?;
-        *self.devices.borrow_mut() = Some(devices.clone());
         Ok(devices)
     }
 }
