+++ streams/device_monitor.rs
@@ -1,0 +1,17 @@
+#[derive(Clone, Copy, Debug)]
+pub enum MediaDeviceKind {
+    AudioInput,
+    AudioOutput,
+    VideoInput,
+}
+
+#[derive(Clone, Debug)]
+pub struct MediaDeviceInfo {
+    pub device_id: String,
+    pub kind: MediaDeviceKind,
+    pub label: String,
+}
+
+pub trait MediaDeviceMonitor {
+    fn enumerate_devices(&self) -> Result<Vec<MediaDeviceInfo>, ()>;
+}
